Separated the input with label from the search component
