+++ src/Search.jsx
@@ -1,4 +1,5 @@
 import PropTypes from "prop-types";
+import InputWithLabel from "./InputWithLabel";
 const Search = ({ search, onSearch }) => {
   const handleChange = (event) => {
     console.log(event.target.value);
@@ -9,11 +10,13 @@ const Search = ({ search, onSearch }) => {
     <>
       <InputWithLabel
         id="search"
-        label="Search"
+        /* label="Search" */
         value={search}
         isFocused
         onInputChange={handleChange}
-      />
+      >
+        Search:
+      </InputWithLabel>
       <p>
         Searching for <strong>{search}</strong>...
       </p>
@@ -21,39 +24,9 @@ const Search = ({ search, onSearch }) => {
   );
 };
 
-const InputWithLabel = ({
-  id,
-  label,
-  value,
-  type = "text",
-  onInputChange,
-  isFocused,
-}) => (
-  <>
-    <label htmlFor={id}>{label}</label>
-    &nbsp;
-    <input
-      id={id}
-      type={type}
-      value={value}
-      autoFocus={isFocused}
-      onChange={onInputChange}
-    />
-  </>
-);
-
 Search.propTypes = {
   onSearch: PropTypes.func.isRequired,
   search: PropTypes.string.isRequired,
-};
-
-InputWithLabel.propTypes = {
-  id: PropTypes.string.isRequired,
-  label: PropTypes.string.isRequired,
-  value: PropTypes.string.isRequired,
-  type: PropTypes.string,
-  isFocused: PropTypes.bool,
-  onInputChange: PropTypes.func.isRequired,
 };
 
 export default Search;
